requirements.txt
python-telegram-bot[job-queue]>=20.0
SQLAlchemy>=2.0.0
APScheduler>=3.10.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
